+++ Assets/Scripts/Utils/SaveData/SaveSingleton.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 82cd17e1a422ec74bb0ce2d547529704
+timeCreated: 1524381632
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
